Fix changed id from int to Long

+++ src/main/java/nl/fontys/realestateproject/business/dto/transaction/MakeTransactionRequest.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MakeTransactionRequest {
-    private int propertyId;
-    private int customerId;
+    private Long propertyId;
+    private Long customerId;
 }
